Refactor hybrid connector into modular package
