+++ src/connectors/hybrid/constants.py
@@ -1,0 +1,4 @@
+"""Constants for the hybrid connector."""
+
+REASONING_PHASE_TIMEOUT = 60.0  # seconds
+EXECUTION_PHASE_TIMEOUT = 120.0  # seconds
